move to jni package
