+++ photon-targeting/src/main/java/org/photonvision/jni/LibraryLoader.java
@@ -27,7 +27,6 @@ import edu.wpi.first.networktables.NetworkTablesJNI;
 import edu.wpi.first.util.WPIUtilJNI;
 import java.io.IOException;
 import org.opencv.core.Core;
-import org.photonvision.common.CombinedRuntimeLoader;
 
 public class LibraryLoader {
     private static boolean hasWpiLoaded = false;
